# Optimize writing of repetition levels in parquet writer