<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description

Avoid iterators and streams and take advantage of Block#mayHaveNull
when writing repetition levels to improve performance.

BenchmarkHiveFileFormat Before
```
write  MAP_VARCHAR_DOUBLE        NONE  TRINO_OPTIMIZED_PARQUET   98.6MB/s ±  1587.5kB/s ( 1.57%) (N = 20, α = 99.9%)
write  LARGE_MAP_VARCHAR_DOUBLE  NONE  TRINO_OPTIMIZED_PARQUET  121.4MB/s ±   548.9kB/s ( 0.44%) (N = 20, α = 99.9%)
write  MAP_INT_DOUBLE            NONE  TRINO_OPTIMIZED_PARQUET  107.8MB/s ±  1137.4kB/s ( 1.03%) (N = 20, α = 99.9%)
write  LARGE_MAP_INT_DOUBLE      NONE  TRINO_OPTIMIZED_PARQUET  110.1MB/s ±   901.0kB/s ( 0.80%) (N = 20, α = 99.9%)
write  LARGE_ARRAY_VARCHAR       NONE  TRINO_OPTIMIZED_PARQUET   91.7MB/s ±  1004.9kB/s ( 1.07%) (N = 20, α = 99.9%)
```

BenchmarkHiveFileFormat After
```
write  MAP_VARCHAR_DOUBLE        NONE  TRINO_OPTIMIZED_PARQUET  214.1MB/s ±  6374.1kB/s ( 2.91%) (N = 20, α = 99.9%)
write  LARGE_MAP_VARCHAR_DOUBLE  NONE  TRINO_OPTIMIZED_PARQUET  247.5MB/s ±  2576.0kB/s ( 1.02%) (N = 20, α = 99.9%)
write  MAP_INT_DOUBLE            NONE  TRINO_OPTIMIZED_PARQUET  319.2MB/s ±  2895.2kB/s ( 0.89%) (N = 20, α = 99.9%)
write  LARGE_MAP_INT_DOUBLE      NONE  TRINO_OPTIMIZED_PARQUET  277.8MB/s ±  4441.1kB/s ( 1.56%) (N = 20, α = 99.9%)
write  LARGE_ARRAY_VARCHAR       NONE  TRINO_OPTIMIZED_PARQUET  167.6MB/s ±  1290.7kB/s ( 0.75%) (N = 20, α = 99.9%)
```

<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues



<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or docs only and no release notes are required.
( ) Release notes are required, please propose a release note for me.
(x) Release notes are required, with the following suggested text:

```markdown
# Hudi, Hive, Delta, Iceberg
* Improve performance of writing structural types in parquet writer. ({issue}`17665`)
```