@raunaqmorarka 2023-05-28
[Optimized parquet writer snappy.pdf](https://github.com/trinodb/trino/files/11585485/Optimized.parquet.writer.snappy.pdf)

<img width="456" alt="Screenshot 2023-05-29 at 12 42 12 AM" src="https://github.com/trinodb/trino/assets/4344846/9d498aeb-f4f8-4eb3-9126-c121da149e00">